move columns to utils
